Remove LSP configuration
Move a couple plugins that I don't think are LSP-related to the editor
plugins module and delete the LSP plugins. I'm not sure I've found LSP
all that useful, and I don't understand the configuration, so I think
it's best to remove it, and if I find a use for it later, I can work out
how to add it all back in.